pyproject.toml
[tool.poetry]
name = "drf-coursework"
version = "0.1.0"
description = ""
authors = ["Tatiana <tanyusha_tokor@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.8"
requests = "^2.32.3"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
coverage = "^7.6.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
